--- domain/src/main/java/com/hous/domain/model/Category.kt
@@ -1,9 +1,0 @@
-package com.hous.domain.model
-
-data class Category(
-    val id: String = "",
-    val categoryName: String = "",
-    val categoryIcon: String = "",
-    val ruleName: String = "",
-    var isChecked: Boolean = false
-)
