--- domain/src/main/java/com/hous/domain/model/DayData.kt
@@ -1,6 +1,0 @@
-package com.hous.domain.model
-
-data class DayData(
-    val day: String,
-    val dayState: State
-)
