--- domain/src/main/java/com/hous/domain/model/Homie.kt
@@ -1,15 +1,0 @@
-package com.hous.domain.model
-
-data class Homie(
-    val id: String? = null,
-    val userName: String = "",
-    val typeName: String = "",
-    val typeId: String = "",
-    val typeColor: String = "",
-    val job: String = "",
-    val introduction: String = "",
-    val hashTag: List<String> = emptyList(),
-    val typeScore: List<Int> = emptyList(),
-    var isChecked: Boolean = false,
-    val notificationState: Boolean = false
-)
